flink-streaming-java-car-topspeed-windowing/data-source/src/main/resources/requirements.txt
avro==1.8.2
kafka-python==0.4.2
